refactor folders structure and remote data source
