--- app/src/main/java/com/picpay/desafio/android/PicPayService.kt
@@ -1,11 +1,0 @@
-package com.picpay.desafio.android
-
-import retrofit2.Call
-import retrofit2.http.GET
-
-
-interface PicPayService {
-
-    @GET("users")
-    fun getUsers(): Call<List<User>>
-}
