+++ app/src/main/java/com/picpay/desafio/android/data/remote/model/User.kt
@@ -1,4 +1,4 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.data.remote.model
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
