+++ app/src/main/java/com/picpay/desafio/android/ui/UserListDiffCallback.kt
@@ -1,7 +1,7 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.ui
 
 import androidx.recyclerview.widget.DiffUtil
-import com.picpay.desafio.android.User
+import com.picpay.desafio.android.data.remote.model.User
 
 class UserListDiffCallback(
     private val oldList: List<User>,
