+++ app/src/main/java/com/picpay/desafio/android/ui/UserListItemViewHolder.kt
@@ -1,7 +1,9 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.ui
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import com.picpay.desafio.android.R
+import com.picpay.desafio.android.data.remote.model.User
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.list_item_user.view.*
