+++ app/src/test/java/com/picpay/desafio/android/ExampleServiceTest.kt
@@ -2,6 +2,8 @@ package com.picpay.desafio.android
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
+import com.picpay.desafio.android.data.remote.data_source.UserRDS
+import com.picpay.desafio.android.data.remote.model.User
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import retrofit2.Call
@@ -9,7 +11,7 @@ import retrofit2.Response
 
 class ExampleServiceTest {
 
-    private val api = mock<PicPayService>()
+    private val api = mock<UserRDS>()
 
     private val service = ExampleService(api)
